Create Yulu - Hypothesis Testing
Yulu is facing a dip in revenues and wants to understand the factors affecting the demand for shared electric cycles in the Indian market. Specifically, Yulu is interested in identifying which variables are significant in predicting the demand for shared electric cycles and how well these variables describe the electric cycle demand.