Create Business Case: Aerofit - Descriptive Statistics & Probability
The market research team at Aerofit wants to identify the characteristics of the target audience for each type of treadmill (KP281, KP481, KP781) to improve recommendations for new customers. They need to perform a detailed analysis to understand customer profiles and how different factors influence the purchase of each treadmill.