Create Business Case: Walmart - Confidence Interval and CLT
Walmart wants to analyze customer purchase behavior, particularly focusing on whether there is a difference in spending habits between male and female customers during Black Friday. The goal is to determine if women spend more than men and use statistical methods to infer this for the entire customer base. The analysis should also be extended to marital status and age groups.